// src/app/models/mycompany.model.ts
// import { FuseUtils } from '@fuse/utils';

export class MyCompany
{
    compaddress: string;
    compafm: string;
    compafm_prefix: string;
    compdoy: string;
    compdrastiriotita: string;
    compemail: string;
    compeponymia: string;
    compfax: string;
    compname: string;
    compphone: string;
    comppoli: string;
    comptk: string;
    compweb: string;
    compworkers: number;
    compxora: string;
    compstate: string;
    compeponymia_eng: string;
    compname_eng: string;
    compdrastiriotita_eng: string;
    compdoy_eng: string;
    comppoli_eng: string;
    compstate_eng: string;
    compxora_eng: string;


  
    // constructor(contact)
    // {
    //     {
    //         this.contid = contact.contid || FuseUtils.generateGUID();
    //         this.contname = contact.contname || '';
    //         this.contsurname = contact.contsurname || '';
    //         this.contphoto = contact.contphoto || 'assets/images/avatars/profile.jpg';
    //         this.contalias = contact.contalias || '';
    //         this.contcompany = contact.contcompany || '';
    //         this.contjobtitle = contact.contjobtitle || '';
    //         this.contmainemail = contact.contmainemail || '';
    //         this.contsecemail = contact.contsecemail || '';
    //         this.contmobile = contact.contmobile || '';
    //         this.contworkphone = contact.contworkphone || '';
    //         this.contsecphone = contact.contsecphone || '';
    //         this.contfax = contact.contfax || '';
    //         this.contaddress = contact.contaddress || '';
    //         this.conttk = contact.conttk || '';
    //         this.contarea = contact.contarea || '';
    //         this.contcity = contact.contcity || '';
    //         this.contcountry = contact.contcountry || '';
    //         this.contdepartment = contact.contdepartment || '';
    //         this.contklados = contact.contklados || '';
    //         this.contlead = contact.contlead || '';
    //         this.contwebsite = contact.contwebsite || '';
    //         this.contdesc = contact.contdesc || '';
    //         this.contbirthday = contact.contbirthday || '';
    //     }
    // }
}